Moved base to views
