+++ packages/flex-plugin-e2e-tests/src/utils/pages/view/agent-desktop.ts
@@ -1,6 +1,6 @@
 import { ElementHandle, Page } from 'puppeteer';
 
-import { Base } from '../base';
+import { Base } from './base';
 
 export class AgentDesktop extends Base {
   private static readonly _noTaskCanvas = '.Twilio-NoTasksCanvas';
