+++ packages/flex-plugin-e2e-tests/src/utils/pages/view/base.ts
